Add support for textalloc in 3D plots
Adds support for allocating text labels in 3D plots. This is done
by adding optional arguments for z, z_scatter and z_lines, and
using the inverse of the projective transform when in 3D mode to
plot the final results.
Also adds a new example for 3D plotting, and adds general updates
to the readme and code formatting.