Add py.typed marker
The textalloc package has full type hints, but many tools will
consider a package to lack type hints unless a `py.typed` marker
is added.

Add the py.typed marker to make it work.

Unsure if anything extra needs to be done to ensure it's picked
up when packaging for PyPI?